requirements.txt
# Window capture
mss>=9.0.0
pywin32>=306

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
websockets>=12.0

# Input
pyautogui>=0.9.54

# Utilities
pydantic>=2.0.0
